requirements.txt
boto3>=1.34.0
botocore>=1.34.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
streamlit>=1.48.0
streamlit-pdf-viewer>=0.1.0
streamlit-folium>=0.15.0
folium>=0.14.0 